Retrieve user profile from database after Sign In
After successful sign in for the first time running the app, it'll check if the profile exists in the database.
If profile found, it means the user have already used the app before so skip the process of setting up the profile (ProfileSetupActivity).
If no reference exists, it means no profile found in the database then the user will be required to set up the profile (i.e. start ProfileSetupActivity).